+++ src/main/resources/reference.conf
@@ -55,10 +55,3 @@ googlecloud {
   priceListEgressKey = "CP-INTERNET-EGRESS-STANDARD-TIER-FROM-NA"
   priceListStorageKey = "CP-BIGSTORE-STORAGE"
 }
-
-trial {
-  spreadsheet {
-    id = ""
-    updateFrequencyMinutes = 0
-  }
-}
