+++ src/main/scala/org/broadinstitute/dsde/firecloud/Application.scala
@@ -19,6 +19,5 @@ case class Application(agoraDAO: AgoraDAO,
                        searchDAO: SearchDAO,
                        researchPurposeSupport: ResearchPurposeSupport,
                        thurloeDAO: ThurloeDAO,
-                       trialDAO: TrialDAO,
                        logitDAO: LogitDAO,
                        shareLogDAO: ShareLogDAO)
