--- src/main/scala/org/broadinstitute/dsde/firecloud/dataaccess/ElasticSearchTrialDAO.scala
@@ -1,388 +1,0 @@
-package org.broadinstitute.dsde.firecloud.dataaccess
-
-import org.broadinstitute.dsde.firecloud.FireCloudException
-import org.broadinstitute.dsde.firecloud.model.ModelJsonProtocol.impTrialProject
-import org.broadinstitute.dsde.firecloud.model.Trial.CreationStatuses.CreationStatus
-import org.broadinstitute.dsde.firecloud.model.WorkbenchUserInfo
-import org.broadinstitute.dsde.firecloud.model.Trial.{CreationStatuses, TrialProject}
-import org.broadinstitute.dsde.rawls.model.RawlsBillingProjectName
-import org.broadinstitute.dsde.workbench.util.health.SubsystemStatus
-import org.elasticsearch.action.admin.indices.create.{CreateIndexRequest, CreateIndexRequestBuilder, CreateIndexResponse}
-import org.elasticsearch.action.admin.indices.exists.indices.{IndicesExistsRequest, IndicesExistsRequestBuilder, IndicesExistsResponse}
-import org.elasticsearch.action.get.{GetRequest, GetRequestBuilder, GetResponse}
-import org.elasticsearch.action.index.{IndexRequest, IndexRequestBuilder, IndexResponse}
-import org.elasticsearch.action.search._
-import org.elasticsearch.action.support.WriteRequest.RefreshPolicy
-import org.elasticsearch.action.update.{UpdateRequest, UpdateRequestBuilder, UpdateResponse}
-import org.elasticsearch.client.transport.TransportClient
-import org.elasticsearch.common.unit.TimeValue
-import org.elasticsearch.common.xcontent.XContentType
-import org.elasticsearch.index.query.QueryBuilder
-import org.elasticsearch.index.query.QueryBuilders._
-import org.elasticsearch.search.sort.SortOrder
-import spray.json._
-
-import scala.concurrent.Future
-import scala.concurrent.ExecutionContext.Implicits.global
-import scala.util.{Failure, Random, Success, Try}
-
-trait TrialQueries {
-
-  val Unverified:QueryBuilder = termQuery("verified", false)
-
-  val Errored:QueryBuilder = boolQuery()
-    .must(termQuery("verified", true))
-    .must(termQuery("status.keyword", CreationStatuses.Error.toString))
-
-  val Available:QueryBuilder = boolQuery()
-    .must(termQuery("verified", true))
-    .must(termQuery("status.keyword", CreationStatuses.Ready.toString))
-    .mustNot(existsQuery("user.userSubjectId.keyword"))
-
-  val Claimed:QueryBuilder = boolQuery()
-    .must(termQuery("verified", true))
-    .must(termQuery("status.keyword", CreationStatuses.Ready.toString))
-    .must(existsQuery("user.userSubjectId.keyword"))
-
-  val Ready:QueryBuilder = boolQuery()
-    .must(termQuery("verified", true))
-    .must(termQuery("status.keyword", CreationStatuses.Ready.toString))
-
-}
-
-/**
-  * DAO, talking to Elasticsearch, for the free trial project pool.
-  *
-  * @param client Elasticsearch client.
-  * @param indexName name of the target free trial index.
-  * @param refreshMode refresh policy. Using IMMEDIATE - the default - is a performance hit but ensures transactionality
-  *                    of updates across multiple users.
-  */
-class ElasticSearchTrialDAO(client: TransportClient, indexName: String, refreshMode: RefreshPolicy = RefreshPolicy.IMMEDIATE)
-  extends TrialDAO with TrialQueries with ElasticSearchDAOSupport {
-
-  lazy private final val datatype = "billingproject"
-
-  init // checks for the presence of the index
-
-  // because the two following methods are not defined at the TrialDAO level, they will
-  // be inaccessible to most callers. Callers should be in the habit of using:
-  //    val dao:TrialDAO = ElasticSearchTrialDAO(...)
-  // which will limit callers to those methods defined in TrialDAO.
-
-  // gets a single project record, with its Elasticsearch version
-  def getProjectInternal(projectName: RawlsBillingProjectName): (Long, TrialProject) = {
-    val getProjectQuery = client.prepareGet(indexName, datatype, projectName.value)
-
-    val getProjectResponse = Try(executeESRequest[GetRequest, GetResponse, GetRequestBuilder](getProjectQuery))
-
-    getProjectResponse match {
-      case Success(get) if get.isExists =>
-        val project = get.getSourceAsString.parseJson.convertTo[TrialProject]
-        val version = get.getVersion
-        (version, project)
-      case Success(notfound) => throw new FireCloudException(s"project ${projectName.value} not found!")
-      case Failure(f) => throw new FireCloudException(s"error retrieving project [${projectName.value}]: ${f.getMessage}")
-    }
-  }
-
-  // update a project, with a check on its Elasticsearch version
-  def updateProjectInternal(updatedProject: TrialProject, version: Long): UpdateResponse = {
-    val update = client
-      .prepareUpdate(indexName, datatype, updatedProject.name.value)
-      .setDoc(updatedProject.toJson.compactPrint, XContentType.JSON)
-      .setVersion(version) // guarantee nobody else has updated in the meantime
-      .setRefreshPolicy(refreshMode)
-
-    executeESRequest[UpdateRequest, UpdateResponse, UpdateRequestBuilder](update) // will throw error if update fails
-  }
-
-  // update a project, with a check on its Elasticsearch version
-  def indexProjectInternal(updatedProject: TrialProject, version: Long): IndexResponse = {
-    val update = client
-      .prepareIndex(indexName, datatype, updatedProject.name.value)
-      .setSource(updatedProject.toJson.compactPrint, XContentType.JSON)
-      .setVersion(version) // guarantee nobody else has updated in the meantime
-      .setRefreshPolicy(refreshMode)
-
-    executeESRequest[IndexRequest, IndexResponse, IndexRequestBuilder](update) // will throw error if update fails
-  }
-
-  /**
-    * Read the record for a specified project. Throws an error if record not found.
-    *
-    * @param projectName name of the project record to read.
-    * @return the project record
-    */
-  override def getProjectRecord(projectName: RawlsBillingProjectName): TrialProject = {
-    val (_, project) = getProjectInternal(projectName)
-    project
-  }
-
-  /**
-    * Check to see if the project record exists.
-    *
-    * @param projectName name of the project record to read.
-    * @return whether or not the project record exists
-    */
-  override def projectRecordExists(projectName: RawlsBillingProjectName): Boolean = {
-    Try(getProjectInternal(projectName)).isSuccess
-  }
-
-  /**
-    * Create a record for the specified project. Throws error if name
-    * already exists or could not be otherwise created.
-    *
-    * @param projectName name of the project to use when creating a record
-    * @return the created project record
-    */
-  override def insertProjectRecord(projectName: RawlsBillingProjectName): TrialProject = {
-    val trialProject = TrialProject(projectName)
-    val insert = client
-      .prepareIndex(indexName, datatype, projectName.value)
-      .setSource(trialProject.toJson.compactPrint, XContentType.JSON)
-      .setCreate(true) // fail the request if the project already exists
-      .setRefreshPolicy(refreshMode)
-
-    executeESRequest[IndexRequest, IndexResponse, IndexRequestBuilder](insert) // will throw error if insert fails
-    trialProject
-  }
-
-  /**
-    * Update the "verified" field for a specified project record. The "verified" field indicates whether
-    * or not the associated billing project was created successfully in Google Cloud. Throws an error if
-    * the record was not found or the record could not be updated.
-    *
-    * @param projectName name of the project record to update
-    * @param verified verified value with which to update the project record
-    * @return the updated project record
-    */
-  override def setProjectRecordVerified(projectName: RawlsBillingProjectName, verified: Boolean, status: CreationStatus): TrialProject = {
-    val (version, project) = getProjectInternal(projectName)
-
-    if (project.verified == verified && project.status.contains(status)) {
-      project // noop
-    } else {
-      val updatedProject = project.copy(verified = verified, status=Some(status))
-      updateProjectInternal(updatedProject, version) // will throw error if update fails
-      updatedProject
-    }
-  }
-
-  /**
-    * Associates the next-available project record with a specified user. "Next available"
-    * is defined as verified, unclaimed, and first alphabetically by project name.
-    * Throws an error if no project records are available, or if the project
-    * record could not be updated.
-    *
-    * @param userInfo the user (email and subjectid) with which to update the project record.
-    * @param randomizationFactor fuzziness for "next-available" - implementation will pick one of the next
-    *                            randomizationFactor projects.
-    * @return the updated project record
-    */
-  override def claimProjectRecord(userInfo: WorkbenchUserInfo, randomizationFactor: Int = 20): TrialProject = {
-    val sliceSize = .5
-
-    // return $randomizationFactor available projects, then choose ($randomizationFactor * $sliceSize) random projects
-    // from that list, and claim the first one we can. Even with this randomization,
-    // the caller of this method should retry this method to ensure a successful claim.
-    val nextProjectRequest = client
-      .prepareSearch(indexName)
-      .setQuery(Available)
-      .addSort("name.keyword", SortOrder.ASC)
-      .setSize(randomizationFactor)
-      .setVersion(true)
-
-    val nextProjectResponse = executeESRequest[SearchRequest, SearchResponse, SearchRequestBuilder](nextProjectRequest)
-
-    if (nextProjectResponse.getHits.totalHits == 0)
-      throw new FireCloudException("Trial has no available projects! Contact a campaign manager to create more.")
-
-    val actualReturned = nextProjectResponse.getHits.getHits.length
-
-    val attemptLimit = Math.ceil(actualReturned * sliceSize)
-
-    def claimInternal(attempt: Int):TrialProject = {
-      val hit = nextProjectResponse.getHits.getAt(Random.nextInt(actualReturned))
-      val version = hit.getVersion
-      val project = hit.getSourceAsString.parseJson.convertTo[TrialProject]
-      assert(project.user.isEmpty)
-
-      val updatedProject = project.copy(user = Some(userInfo))
-      Try(updateProjectInternal(updatedProject, version)) match {
-        case Success(_) => updatedProject
-        case Failure(f) =>
-          if (attempt > attemptLimit) {
-            throw f
-          } else {
-            logger.debug(s"claimProjectRecord retrying claim; attempt $attempt")
-            claimInternal(attempt+1)
-          }
-      }
-    }
-
-    claimInternal(0)
-  }
-
-  /**
-    * Removes the associated user from the project record, making it available again.
-    * Throws an error if the project record could not be updated.
-    *
-    * @param projectName the name of the project to return to the available pool
-    */
-  override def releaseProjectRecord(projectName: RawlsBillingProjectName): TrialProject = {
-    val (version, project) = getProjectInternal(projectName)
-    project.user match {
-      case Some(_) => {
-        val updatedProject = project.copy(user = None)
-        indexProjectInternal(project.copy(user = None), version)
-        updatedProject
-      }
-      case None => project
-    }
-  }
-
-  /**
-    * Returns a list of project records in the pool that are unverified.
-    *
-    * @return list of project records in the pool that are unverified.
-    */
-  override def listUnverifiedProjects: Seq[TrialProject] = {
-    val unverifiedRequest = client
-      .prepareSearch(indexName)
-      .setQuery(Unverified)
-      .setSize(1000)
-
-    val unverifiedResponse = executeESRequest[SearchRequest, SearchResponse, SearchRequestBuilder](unverifiedRequest)
-
-    if (unverifiedResponse.getHits.totalHits == 0)
-      Seq.empty[TrialProject]
-    else
-      unverifiedResponse.getHits.getHits.toSeq map ( _.getSourceAsString.parseJson.convertTo[TrialProject] )
-  }
-
-  /**
-    * Returns a counts of project records by status.
-    * @return counts of project records.
-    */
-  override def countProjects: Map[String,Long] = {
-    Map(
-      "unverified" -> count(Unverified),
-      "errored" -> count(Errored),
-      "available" -> count(Available),
-      "claimed" -> count(Claimed)
-    )
-  }
-
-  private def handleScrollResponse(scrollResponse: SearchResponse): (String, List[TrialProject]) = {
-    val scrollId: String = scrollResponse.getScrollId
-
-    val trialProjects: List[TrialProject] = if (scrollResponse.getHits.totalHits == 0) {
-      List.empty[TrialProject]
-    } else {
-      scrollResponse.getHits.getHits.toList map ( _.getSourceAsString.parseJson.convertTo[TrialProject] )
-    }
-
-    (scrollId, trialProjects)
-  }
-
-  private def startScroll: (String, List[TrialProject]) = {
-    /* Start an Elasticsearch scroll request. This allows us to retrieve all records in the index,
-        in batches of N records.
-
-        We currently have N set to 250.
-
-        Dear future engineer: if you find that you ever need to change this value, even
-        just to experiment with other values, or to use different values in different environments,
-        please move it to config so that it is easier to tweak.
-    */
-    val startScrollRequest = client
-      .prepareSearch(indexName)
-      .setQuery(Ready)
-      .setScroll(TimeValue.timeValueMinutes(1))
-      .addSort("_doc", SortOrder.ASC)
-      .setSize(250) // tweak pagesize here
-
-    handleScrollResponse(
-      executeESRequest[SearchRequest, SearchResponse, SearchRequestBuilder](startScrollRequest))
-  }
-
-  private def continueScroll(scrollId: String): (String, List[TrialProject]) = {
-    val continueScrollRequest = client
-      .prepareSearchScroll(scrollId)
-      .setScroll(TimeValue.timeValueMinutes(1))
-
-    handleScrollResponse(
-      executeESRequest[SearchScrollRequest, SearchResponse, SearchScrollRequestBuilder](continueScrollRequest))
-  }
-
-  private def clearScroll(scrollId: String): Future[Any] = {
-    // we want to clear the scroll to release resources on the ES server.
-    // however, ES will automatically clean up scrolls after their timeout.
-    // so, we make this a fire-and-forget request, and don't care whether or not it succeeds or fails.
-    val clearScrollRequest = client
-      .prepareClearScroll()
-      .addScrollId(scrollId)
-    Future(executeESRequest[ClearScrollRequest, ClearScrollResponse, ClearScrollRequestBuilder](clearScrollRequest))
-  }
-
-  /**
-    * Returns a list of all project records.
-    * @return list of all project records.
-    */
-  override def projectReport: Seq[TrialProject] = {
-
-    def nextScroll(scrollId: Option[String], accum: List[TrialProject]): (String, List[TrialProject]) = {
-      // is this the initial query?
-      val (newScrollId, batch) = scrollId match {
-        case None => startScroll
-        case Some(s) => continueScroll(s)
-      }
-
-      // if the query returned results, add them to the accumulator and make another query
-      if (batch.nonEmpty) {
-        nextScroll(Some(newScrollId), accum ++ batch)
-      } else {
-        // else, release the scroll in ES and return results
-        clearScroll(newScrollId)
-        (newScrollId, accum)
-      }
-    }
-
-    val (_, finalAccum) = nextScroll(None, List.empty[TrialProject])
-
-    finalAccum.sortBy(_.name.value)
-  }
-
-  private def indexExists: Boolean = {
-    executeESRequest[IndicesExistsRequest, IndicesExistsResponse, IndicesExistsRequestBuilder](
-      client.admin.indices.prepareExists(indexName)
-    ).isExists
-  }
-
-  override def status: Future[SubsystemStatus] = {
-    Future(SubsystemStatus(indexExists, None))
-  }
-
-  private def init: Unit = {
-    if (!indexExists) {
-      executeESRequest[CreateIndexRequest, CreateIndexResponse, CreateIndexRequestBuilder](
-        client.admin.indices.prepareCreate(indexName))
-      // Try one more time and fail if index creation fails
-      if (!indexExists)
-        throw new FireCloudException(s"index $indexName does not exist!")
-    }
-  }
-
-  private def count(qb: QueryBuilder): Long = {
-    val countProjectRequest = client
-      .prepareSearch(indexName)
-      .setQuery(qb)
-      .setSize(0)
-
-    val countProjectResponse = executeESRequest[SearchRequest, SearchResponse, SearchRequestBuilder](countProjectRequest)
-
-    countProjectResponse.getHits.totalHits
-  }
-
-}
