+++ src/main/scala/org/broadinstitute/dsde/firecloud/dataaccess/HttpThurloeDAO.scala
@@ -2,7 +2,6 @@ package org.broadinstitute.dsde.firecloud.dataaccess
 
 import akka.actor.ActorSystem
 import org.broadinstitute.dsde.firecloud.model.ModelJsonProtocol._
-import org.broadinstitute.dsde.firecloud.model.Trial.UserTrialStatus
 import org.broadinstitute.dsde.firecloud.model._
 import org.broadinstitute.dsde.firecloud.utils.RestJsonClient
 import org.broadinstitute.dsde.firecloud.webservice.UserApiService
@@ -79,36 +78,6 @@ class HttpThurloeDAO ( implicit val system: ActorSystem, implicit val executionC
     val profilePropertyMap = profile.propertyValueMap
     saveKeyValues(userInfo, profilePropertyMap).map(_ => ())
   }
-  /**
-    * get the UserTrialStatus associated with a specific user.
-    *
-    * @param forUserId the subjectid of the user whose trial status to get
-    * @param callerToken the OAuth token of the person making the API call
-    * @return the trial status for the specified user, or a default UserTrialStatus object
-    */
-  override def getTrialStatus(forUserId: String, callerToken: WithAccessToken): Future[UserTrialStatus] = {
-    getAllKVPs(forUserId, callerToken) map {
-      case Some(wrapper) =>
-        val status = UserTrialStatus(wrapper)
-        assert(forUserId == status.userId, "status id does not match!")
-        status
-      case None =>
-        throw new FireCloudException("Unable to get user trial status")
-    }
-  }
-
-  /**
-    * set the UserTrialStatus for a specific user
-    *
-    * @param forUserId the subjectid of the user whose trial status to set
-    * @param callerToken the OAuth token of the person making the API call
-    * @param trialStatus the trial status to save for the specified user
-    * @return success/failure of whether or not the status saved correctly
-    */
-  override def saveTrialStatus(forUserId: String, callerToken: WithAccessToken, trialStatus: UserTrialStatus): Future[Try[Unit]] = {
-    saveKeyValues(forUserId, callerToken, UserTrialStatus.toKVPs(trialStatus))
-  }
-
 
   private def wrapExceptions[T](codeBlock: => Future[T]): Future[T] = {
     codeBlock.recover {
