+++ src/main/scala/org/broadinstitute/dsde/firecloud/dataaccess/ThurloeDAO.scala
@@ -1,7 +1,6 @@
 package org.broadinstitute.dsde.firecloud.dataaccess
 
 import com.typesafe.scalalogging.LazyLogging
-import org.broadinstitute.dsde.firecloud.model.Trial.UserTrialStatus
 import org.broadinstitute.dsde.firecloud.model.{BasicProfile, ProfileWrapper, UserInfo, WithAccessToken}
 import org.broadinstitute.dsde.rawls.model.ErrorReportSource
 
@@ -40,27 +39,6 @@ trait ThurloeDAO extends LazyLogging with ReportsSubsystemStatus {
     * @return success/failure of save
     */
   def saveKeyValues(forUserId: String, callerToken: WithAccessToken, keyValues: Map[String, String]): Future[Try[Unit]]
-
-  // methods to work with free-trial objects
-  /**
-    * get the UserTrialStatus associated with a specific user or a default UserTrialStatus object.
-    *
-    * @param forUserId the subjectid of the user whose trial status to get
-    * @param callerToken the OAuth token of the person making the API call
-    * @return the trial status for the specified user, or a default UserTrialStatus object
-
-    */
-  def getTrialStatus(forUserId: String, callerToken: WithAccessToken): Future[UserTrialStatus]
-
-  /**
-    * set the UserTrialStatus for a specific user
-    *
-    * @param forUserId the subjectid of the user whose trial status to set
-    * @param callerToken the OAuth token of the person making the API call
-    * @param trialStatus the trial status to save for the specified user
-    * @return success/failure of whether or not the status saved correctly
-    */
-  def saveTrialStatus(forUserId: String, callerToken: WithAccessToken, trialStatus: UserTrialStatus): Future[Try[Unit]]
 
   def bulkUserQuery(userIds: List[String], keySelection: List[String]): Future[List[ProfileWrapper]]
