+++ src/main/scala/org/broadinstitute/dsde/firecloud/utils/PermissionsSupport.scala
@@ -1,13 +1,13 @@
 package org.broadinstitute.dsde.firecloud.utils
 
 import org.broadinstitute.dsde.firecloud.dataaccess.{RawlsDAO, SamDAO}
-import org.broadinstitute.dsde.firecloud.{FireCloudConfig, FireCloudException, FireCloudExceptionWithErrorReport}
 import org.broadinstitute.dsde.firecloud.model._
-import org.broadinstitute.dsde.rawls.model.{ErrorReport, RawlsGroupName}
 import org.broadinstitute.dsde.firecloud.service.PerRequest.PerRequestMessage
+import org.broadinstitute.dsde.firecloud.{FireCloudException, FireCloudExceptionWithErrorReport}
+import org.broadinstitute.dsde.rawls.model.ErrorReport
 import org.broadinstitute.dsde.rawls.model.WorkspaceAccessLevels.WorkspaceAccessLevel
 import org.broadinstitute.dsde.workbench.model.WorkbenchGroupName
-import spray.http.{HttpRequest, StatusCodes}
+import spray.http.StatusCodes
 
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -56,9 +56,6 @@ trait PermissionsSupport {
       }
     }
   }
-
-  def asTrialCampaignManager(op: => Future[PerRequestMessage])(implicit userInfo: UserInfo): Future[PerRequestMessage] =
-    asGroupMember(FireCloudConfig.Trial.managerGroup)(op)
 
   def asGroupMember(group: String)(op: => Future[PerRequestMessage])(implicit userInfo: UserInfo): Future[PerRequestMessage] = {
     tryIsGroupMember(userInfo, group) flatMap { isGroupMember =>
