+++ src/main/scala/org/broadinstitute/dsde/firecloud/webservice/UserApiService.scala
@@ -63,7 +63,6 @@ trait UserApiService extends HttpService with PerRequestCreator with FireCloudRe
 
   lazy val log = LoggerFactory.getLogger(getClass)
 
-  val trialServiceConstructor: () => TrialService
   val userServiceConstructor: (UserInfo) => UserService
 
   val userServiceRoutes =
@@ -134,35 +133,6 @@ trait UserApiService extends HttpService with PerRequestCreator with FireCloudRe
         delete {
           requireUserInfo() { userInfo => requestContext =>
             perRequest(requestContext, UserService.props(userServiceConstructor, userInfo), DeleteTerraPreference)
-          }
-        }
-      } ~
-      pathPrefix("profile" / "trial") {
-        pathEnd {
-          post {
-            parameter("operation" ? "enroll") { op =>
-              requireUserInfo() { userInfo => requestContext =>
-                val operation = op.toLowerCase match {
-                  case "enroll" => Some(TrialService.EnrollUser(userInfo))
-                  case "finalize" => Some(TrialService.FinalizeUser(userInfo))
-                  case _ => None
-                }
-
-                if (operation.nonEmpty)
-                  perRequest(requestContext, TrialService.props(trialServiceConstructor), operation.get)
-                else
-                  requestContext.complete(BadRequest, ErrorReport(s"Invalid operation '$op'"))
-              }
-            }
-          }
-        } ~
-        path("userAgreement") {
-          put {
-            requireUserInfo() { userInfo => requestContext =>
-              perRequest(requestContext,
-                TrialService.props(trialServiceConstructor),
-                TrialService.RecordUserAgreement(userInfo))
-            }
           }
         }
       } ~
