+++ src/test/resources/reference.conf
@@ -12,8 +12,6 @@ auth {
   jsonFile = "/dev/null"
   rawlsPemFile = "/dev/null"
   rawlsPemFileClientId = "dummy"
-  trialBillingPemFile = "/dev/null"
-  trialBillingPemFileClientId = "dummy"
 
   swaggerRealm = "broad-dsde-dev"
 }
@@ -75,7 +73,6 @@ elasticsearch {
   clusterName = "elasticsearch5a"
   index = "unittest"
   ontologyIndex = "ontology-unittest"
-  trialIndex = "trial-unittest"
   shareLogIndex = "sharelog-unittest"
   discoverGroupNames = ["all_broad_users","demo_users"]
 }
@@ -84,13 +81,6 @@ duos {
   baseConsentUrl = "http://localhost:8992"
   baseOntologyUrl = "http://localhost:8993"
   dulvn = 4
-}
-
-trial {
-  durationDays = 60
-  managerGroup = "trial_managers"
-  billingAccount = "billingAccounts/dummy"
-  projectBufferSize = 2
 }
 
 metrics {
