+++ src/test/scala/org/broadinstitute/dsde/firecloud/service/BaseServiceSpec.scala
@@ -16,11 +16,10 @@ class BaseServiceSpec extends ServiceSpec with BeforeAndAfter {
   val searchDao:MockSearchDAO = new MockSearchDAO
   val researchPurposeSupport:MockResearchPurposeSupport = new MockResearchPurposeSupport
   val thurloeDao:MockThurloeDAO = new MockThurloeDAO
-  val trialDao:MockTrialDAO = new MockTrialDAO
   val logitDao:MockLogitDAO = new MockLogitDAO
   val shareLogDao:MockShareLogDAO = new MockShareLogDAO
 
   val app:Application =
-    new Application(agoraDao, googleServicesDao, ontologyDao, consentDao, rawlsDao, samDao, searchDao, researchPurposeSupport, thurloeDao, trialDao, logitDao, shareLogDao)
+    new Application(agoraDao, googleServicesDao, ontologyDao, consentDao, rawlsDao, samDao, searchDao, researchPurposeSupport, thurloeDao, logitDao, shareLogDao)
 
 }
