--- src/test/scala/org/broadinstitute/dsde/firecloud/trial/ProjectManagerSpec.scala
@@ -1,348 +1,0 @@
-package org.broadinstitute.dsde.firecloud.trial
-
-import akka.actor.{Actor, ActorRef, ActorSystem, Props}
-import akka.pattern.ask
-import akka.testkit.TestKit
-import akka.util.Timeout
-import org.broadinstitute.dsde.firecloud.FireCloudException
-import org.broadinstitute.dsde.firecloud.dataaccess._
-import org.broadinstitute.dsde.firecloud.mock.MockGoogleServicesDAO
-import org.broadinstitute.dsde.firecloud.model.Trial.{CreationStatuses, ProjectRoles, RawlsBillingProjectMembership, TrialProject}
-import org.broadinstitute.dsde.firecloud.model.{Trial, WithAccessToken, WorkbenchUserInfo}
-import org.broadinstitute.dsde.firecloud.trial.ProjectManager._
-import org.broadinstitute.dsde.rawls.model.RawlsBillingProjectName
-import org.scalatest.{BeforeAndAfterAll, FreeSpecLike}
-
-import scala.concurrent.duration._
-import scala.concurrent.{Await, Future}
-import scala.language.postfixOps
-
-class ProjectManagerSpec extends TestKit(ActorSystem("ProjectManagerSpec")) with FreeSpecLike with BeforeAndAfterAll {
-
-  import ProjectManagerSpec._
-
-  override def afterAll {
-    TestKit.shutdownActorSystem(system)
-  }
-
-  implicit val askTimeout: Timeout = Timeout(5.seconds) // timeout for getting a response from the ProjectManager actor
-  val testTimeout: FiniteDuration = 10.seconds // timeout for a test to succeed or fail when using awaitCond/awaitAssert
-  val testInterval: FiniteDuration = 150.millis // interval on which to repeatedly re-check a test when using awaitCond/awaitAssert
-
-  val rawlsDAO = new ProjectManagerSpecRawlsDAO
-  val trialDAO = new ProjectManagerSpecTrialDAO
-  val googleDAO = new MockGoogleServicesDAO
-
-  /*
-    call stack for project manager:
-      - trialDAO.insertProjectRecord
-      - rawlsDAO.createProject
-      - rawlsDAO.getProjects
-      - trialDAO.setProjectRecordVerified
-   */
-
-  "ProjectManager actor, when creating projects" - {
-    "should insert a record" in {
-      val (pm, rawlsDAO, trialDAO, googleDAO) = initTestState()
-      pm ! StartCreation(1)
-      awaitCond(trialDAO.insertCount == 1, testTimeout, testInterval)
-      system.stop(pm)
-    }
-    "should insert multiple records" in {
-      val (pm, rawlsDAO, trialDAO, googleDAO) = initTestState()
-      pm ! StartCreation(5)
-      awaitCond(trialDAO.insertCount == 5, testTimeout, testInterval)
-      system.stop(pm)
-    }
-    "multiple inserted records should all be unique" in {
-      val (pm, rawlsDAO, trialDAO, googleDAO) = initTestState()
-      pm ! StartCreation(5)
-      awaitCond(insertedProjects.map(_.name).distinct.size == 5, testTimeout, testInterval)
-      system.stop(pm)
-    }
-
-    "should impose a delay between multiple project creations" in {
-      // with a createDelay of 5 seconds, creating 5 project should take at least 25 seconds
-      // previous test proves that we can create 5 projects (in the mocks!) within 3 seconds
-      val (pm, rawlsDAO, trialDAO, googleDAO) = initTestState(createDelay = 5.seconds)
-      pm ! StartCreation(5)
-      Thread.sleep(3000)
-      assert(trialDAO.insertCount < 5)
-      system.stop(pm)
-    }
-
-    "should trigger project creation in rawls" in {
-      val (pm, rawlsDAO, trialDAO, googleDAO) = initTestState()
-      pm ! StartCreation(5)
-      awaitCond(trialDAO.insertCount == 5, testTimeout, testInterval)
-      awaitCond(rawlsDAO.createCount == 5, testTimeout, testInterval)
-      system.stop(pm)
-    }
-
-    "should trigger verification of inserted records" in {
-      val (pm, rawlsDAO, trialDAO, googleDAO) = initTestState()
-      pm ! StartCreation(5)
-      awaitCond(trialDAO.insertCount == 5 && insertedProjects.map(_.name).toSet == verifiedProjects.map(_.name).toSet, testTimeout, testInterval)
-      system.stop(pm)
-    }
-    "should impose a delay between project creation and verification" in {
-      // with a verify delay of 5 seconds and a sleep of 3, we should see no verifications.
-      // previous test proves that verifications are triggered correctly, this test checks the delay.
-      val (pm, rawlsDAO, trialDAO, googleDAO) = initTestState(verifyDelay = 5.seconds)
-      pm ! StartCreation(5)
-      Thread.sleep(3000)
-      assert(trialDAO.verifiedCount == 0)
-      system.stop(pm)
-    }
-    "should only verify a project once if it verifies on the first try" in {
-      // we use the rawls getProjects() call as a proxy for verification - in ProjectManager,
-      // getProjects() is only called within the verify handler.
-      val (pm, rawlsDAO, trialDAO, googleDAO) = initTestState()
-      pm ! StartCreation(1)
-      Thread.sleep(3000)
-      // under normal conditions, verification succeeds and only executes once.
-      assert(rawlsGetProjectsCallCount == 1)
-      system.stop(pm)
-    }
-    "should re-verify a project if it is still creating" in {
-      // we use the rawls getProjects() call as a proxy for verification - in ProjectManager,
-      // getProjects() is only called within the verify handler.
-      val (pm, rawlsDAO, trialDAO, googleDAO) = initTestState(mockRawlsDAO = new StillCreatingProjectsRawlsDAO)
-      pm ! StartCreation(1)
-      Thread.sleep(3000)
-      // if projects are still creating, and we have zero verification delay, rawlsGetProjectsCallCount should
-      // be much much higher than 1.
-      assert(rawlsGetProjectsCallCount > 1 && verifiedProjects.isEmpty)
-      system.stop(pm)
-    }
-    s"should only retry verification ${ProjectManager.verificationLimit} times" in {
-      // we use the rawls getProjects() call as a proxy for verification - in ProjectManager,
-      // getProjects() is only called within the verify handler.
-      val (pm, rawlsDAO, trialDAO, googleDAO) = initTestState(mockRawlsDAO = new StillCreatingProjectsRawlsDAO)
-      pm ! StartCreation(1)
-      Thread.sleep(3000)
-      // if projects are still creating, and we have zero verification delay, we should be able to reach
-      // the limit within 3 seconds.
-      assert(rawlsGetProjectsCallCount == ProjectManager.verificationLimit && verifiedProjects.isEmpty)
-      system.stop(pm)
-    }
-    "should mark projects as error during verification if they failed to create" in {
-      val (pm, rawlsDAO, trialDAO, googleDAO) = initTestState(mockRawlsDAO = new BadGetProjectsRawlsDAO)
-      pm ! StartCreation(3)
-      awaitCond(trialDAO.insertCount == 3, testTimeout, testInterval)
-      awaitCond(trialDAO.verifiedCount == 3 && verifiedProjects.forall(p => p.verified && p.status.contains(CreationStatuses.Error)), testTimeout, testInterval)
-      system.stop(pm)
-    }
-    "should mark projects as error during verification if they are not returned by Rawls" in {
-      val (pm, rawlsDAO, trialDAO, googleDAO) = initTestState(mockRawlsDAO = new MissingGetProjectsRawlsDAO)
-      pm ! StartCreation(3)
-      awaitCond(trialDAO.insertCount == 3, testTimeout, testInterval)
-      awaitCond(trialDAO.verifiedCount == 3 && verifiedProjects.forall(p => p.verified && p.status.contains(CreationStatuses.Error)), testTimeout, testInterval)
-      system.stop(pm)
-    }
-
-  }
-
-
-  private def initTestState(mockRawlsDAO: ProjectManagerSpecRawlsDAO = new ProjectManagerSpecRawlsDAO,
-                            mockTrialDAO: ProjectManagerSpecTrialDAO = new ProjectManagerSpecTrialDAO,
-                            mockGoogleDAO: ProjectManagerSpecMockGoogleServicesDAO = new ProjectManagerSpecMockGoogleServicesDAO,
-                            createDelay: FiniteDuration = Duration.Zero,
-                            verifyDelay: FiniteDuration = Duration.Zero):
-  (ActorRef, ProjectManagerSpecRawlsDAO, ProjectManagerSpecTrialDAO, ProjectManagerSpecMockGoogleServicesDAO) = {
-
-    initVars
-
-    // start a project manager, with zero delays for creating projects and verifying projects
-    val pm = system.actorOf(
-      ProjectManager.props(mockRawlsDAO, mockTrialDAO, mockGoogleDAO, createDelay, verifyDelay),
-      java.util.UUID.randomUUID().toString)
-
-    (pm, mockRawlsDAO, mockTrialDAO, mockGoogleDAO)
-  }
-
-}
-
-
-object ProjectManagerSpec {
-
-  // ****************************************************
-  // mutable shared state variables used by the mock DAOs
-  import MockDAOData._
-
-  def initVars(): Unit = {
-    projectStateActor ! Init
-  }
-
-  def verifiedProjects: Seq[TrialProject] = {
-    val f = ask(projectStateActor, GetVerifiedProjects).mapTo[Seq[TrialProject]]
-    Await.result(f, timeout.duration)
-  }
-
-  def insertedProjects: Seq[TrialProject] = {
-    val f = ask(projectStateActor, GetInsertedProjects).mapTo[Seq[TrialProject]]
-    Await.result(f, timeout.duration)
-  }
-
-  def rawlsCreatedProjects: Seq[String] = {
-    val f = ask(projectStateActor, GetCreatedProjects).mapTo[Seq[String]]
-    Await.result(f, timeout.duration)
-  }
-
-  def increaseRawlsProjectsCallCount(): Unit = {
-    projectStateActor ! BumpRawlsProjectCallsCount
-  }
-
-  def rawlsGetProjectsCallCount: Int = {
-    val f = ask(projectStateActor, GetProjectCallsCount).mapTo[Int]
-    Await.result(f, timeout.duration)
-  }
-
-  /**
-    * Rawls dao with test instrumentation
-    */
-  class ProjectManagerSpecRawlsDAO extends MockRawlsDAO {
-
-    def createCount: Int = rawlsCreatedProjects.size
-
-    override def createProject(projectName: String, billingAccount: String)(implicit userToken: WithAccessToken): Future[Boolean] = {
-      projectStateActor ! RawlsCreatedProjects(projectName)
-      Future.successful(true)
-    }
-
-    override def getProjects(implicit userToken: WithAccessToken): Future[Seq[Trial.RawlsBillingProjectMembership]] = {
-      increaseRawlsProjectsCallCount
-      val projects = rawlsCreatedProjects map { name =>
-        RawlsBillingProjectMembership(RawlsBillingProjectName(name), ProjectRoles.Owner, CreationStatuses.Ready, None)
-      }
-      Future.successful(projects)
-    }
-  }
-
-  /**
-    * Rawls dao that fails to return projects in the getProjects query
-    */
-  class MissingGetProjectsRawlsDAO extends ProjectManagerSpecRawlsDAO {
-    override def getProjects(implicit userToken: WithAccessToken): Future[Seq[Trial.RawlsBillingProjectMembership]] = {
-      increaseRawlsProjectsCallCount
-      Future.successful(Seq.empty[RawlsBillingProjectMembership])
-    }
-  }
-
-  /**
-    * Rawls dao that says that projects failed to create in Google land
-    */
-  class BadGetProjectsRawlsDAO extends ProjectManagerSpecRawlsDAO {
-    override def getProjects(implicit userToken: WithAccessToken): Future[Seq[Trial.RawlsBillingProjectMembership]] = {
-      increaseRawlsProjectsCallCount
-      val projects = rawlsCreatedProjects map { name =>
-        RawlsBillingProjectMembership(RawlsBillingProjectName(name), ProjectRoles.Owner, CreationStatuses.Error, Some("unit test creation error"))
-      }
-      Future.successful(projects)
-    }
-  }
-
-  /**
-    * Rawls dao that says that projects are still creating in Google land
-    */
-  class StillCreatingProjectsRawlsDAO extends ProjectManagerSpecRawlsDAO {
-    override def getProjects(implicit userToken: WithAccessToken): Future[Seq[Trial.RawlsBillingProjectMembership]] = {
-      increaseRawlsProjectsCallCount
-      val projects = rawlsCreatedProjects map { name =>
-        RawlsBillingProjectMembership(RawlsBillingProjectName(name), ProjectRoles.Owner, CreationStatuses.Creating, None)
-      }
-      Future.successful(projects)
-    }
-  }
-
-  class ProjectManagerSpecMockGoogleServicesDAO extends MockGoogleServicesDAO
-
-
-  class ProjectManagerSpecTrialDAO extends MockTrialDAO {
-
-    def insertCount(): Int = insertedProjects.size
-
-    def verifiedCount: Int = verifiedProjects.size
-
-    override def insertProjectRecord(projectName: RawlsBillingProjectName): TrialProject = {
-      if (insertedProjects.exists(p => p.name == projectName)) {
-        throw new FireCloudException("ProjectManagerSpecTrialDAO says not unique")
-      } else {
-        val insertedProject = TrialProject(projectName)
-        projectStateActor ! InsertProject(insertedProject)
-        insertedProject
-      }
-    }
-
-    override def getProjectRecord(projectName: RawlsBillingProjectName): TrialProject = insertedProjects.reverse.head
-
-    override def projectRecordExists(projectName: RawlsBillingProjectName): Boolean = true
-
-    override def setProjectRecordVerified(projectName: RawlsBillingProjectName, verified: Boolean, status: CreationStatuses.CreationStatus): TrialProject = {
-      val verifiedProject = TrialProject(projectName, verified, None, Some(status))
-      projectStateActor ! VerifiedProject(verifiedProject)
-      verifiedProject
-    }
-
-    override def claimProjectRecord(userInfo: WorkbenchUserInfo, randomizationFactor: Int = 20): TrialProject =
-      TrialProject(RawlsBillingProjectName("unittest"))
-
-    override def releaseProjectRecord(projectName: RawlsBillingProjectName): TrialProject =
-      TrialProject(projectName)
-
-    override def listUnverifiedProjects: Seq[TrialProject] = Seq.empty[TrialProject]
-
-    override def countProjects: Map[String, Long] = Map.empty[String, Long]
-
-    override def projectReport: Seq[TrialProject] = Seq.empty[TrialProject]
-  }
-
-}
-
-object MockDAOData {
-
-  val system = ActorSystem("MockDAOData")
-  val projectStateActor: ActorRef = system.actorOf(Props[RegisterTrialProjectActor], name = "MockDAOData")
-  implicit val timeout: Timeout = Timeout(3.seconds)
-
-  case object Init
-
-  // Setters
-  case class InsertProject(p: TrialProject)
-  case class VerifiedProject(p: TrialProject)
-  case object BumpRawlsProjectCallsCount
-  case class RawlsCreatedProjects(p: String)
-
-  // Getters
-  case object GetInsertedProjects
-  case object GetVerifiedProjects
-  case object GetProjectCallsCount
-  case object GetCreatedProjects
-
-  //noinspection ActorMutableStateInspection
-  class RegisterTrialProjectActor extends Actor {
-
-    private var insertedProjects = Seq.empty[TrialProject]
-    private var verifiedProjects = Seq.empty[TrialProject]
-    private var rawlsProjectsCallCount = 0
-    private var rawlsCreatedProjects = Seq.empty[String]
-
-    override def receive: Receive = {
-      case Init =>
-        insertedProjects = Seq.empty[TrialProject]
-        verifiedProjects = Seq.empty[TrialProject]
-        rawlsProjectsCallCount = 0
-        rawlsCreatedProjects = Seq.empty[String]
-      case InsertProject(project: TrialProject) => insertedProjects = insertedProjects :+ project
-      case VerifiedProject(project: TrialProject) => verifiedProjects = verifiedProjects :+ project
-      case BumpRawlsProjectCallsCount => rawlsProjectsCallCount += 1
-      case RawlsCreatedProjects(project: String) => rawlsCreatedProjects = rawlsCreatedProjects :+ project
-      case GetInsertedProjects => sender ! insertedProjects
-      case GetVerifiedProjects => sender ! verifiedProjects
-      case GetProjectCallsCount => sender ! rawlsProjectsCallCount
-      case GetCreatedProjects => sender ! rawlsCreatedProjects
-    }
-
-  }
-
-}
